Move merge institutions and merge users files to lib/tasks
