+++ lib/tasks/merge_users.rb
